Create "Sort JSON Array.js"
An interview question that asked me to sort a JSON-array with articles in ascending order based on the "inventory" value. If the value is the same then the newer item should come first.

This first commit has finished most of the above. The user inputs an array and it then prints out in the terminal the sorted order. Use a JSON prettifier to make reading easier!