Now sorts based on date if inventory is equal
The task also specificed to sort the objects based on date (earliest to latest) if the values of inventory happen to be equal. It now also does that after doing a little fix regarding the quotation marks that were used. See comments in code. Added a ton of documentation too. Code can probably be cleaner but this is a very detailed and step-by-step implementation showing every part of the way that I've taken to solve this.