Input Validation function finished
Use the .html file to try out the function itself. The .js file is just the function, but takes in an Array as parameter instead. The .html file has a simple UI and input form that the user can fill out to test out the function. The values that are tested will be whatever the user fills out.